refactor(app): extract attachment viewer handling from inputs
