+++ src/app/inputs.rs
@@ -15,7 +15,6 @@ use crate::app::input_mapping::{
 use crate::app::status_input::status_view_allows;
 use crate::file_picker::FilePickerState;
 use crate::key_handler::Key;
-use std::path::Path;
 use whatsrust as wr;
 
 impl App<'_> {
@@ -430,183 +429,6 @@ impl App<'_> {
     fn cancel_file_picker(&mut self) {
         self.file_picker = None;
         self.action_notice = Some(crate::app::actions::ActionNotice::Cancelled);
-    }
-
-    fn open_attachment_viewer(&mut self) {
-        if self.selected_message_is_deleted() {
-            return self.unavailable("This message was deleted.");
-        }
-        let Some(selected) = self.selected_message().cloned() else {
-            return self.unavailable("View is not available");
-        };
-        let wr::MessageContent::File(file) = &selected.message else {
-            return self.unavailable("View is not available");
-        };
-        if matches!(file.kind, wr::FileKind::Sticker) {
-            self.action_notice = Some(crate::app::actions::ActionNotice::Unsupported(
-                "Sticker viewer is not supported".into(),
-            ));
-            return;
-        }
-        if !matches!(file.kind, wr::FileKind::Image | wr::FileKind::Video) {
-            return self.unavailable("View is not available");
-        }
-        // In the Status section the viewer navigates only the opened
-        // contact's statuses; everywhere else it uses the open chat's media.
-        let pool = if self.selected_section == Section::Status {
-            self.open_status_contact()
-                .map(|contact| self.status_messages(&contact))
-                .unwrap_or_default()
-        } else {
-            self.chat_messages
-                .get(&selected.info.chat)
-                .cloned()
-                .unwrap_or_default()
-        };
-        let mut attachments = pool
-            .iter()
-            .filter_map(|id| self.messages.get(id))
-            .filter_map(|message| match &message.message {
-                wr::MessageContent::File(file)
-                    if matches!(file.kind, wr::FileKind::Image | wr::FileKind::Video) =>
-                {
-                    Some(crate::app::events::ViewerAttachment {
-                        message_id: message.info.id.clone(),
-                        kind: file.kind.clone(),
-                        path: file.path.clone(),
-                        status: self.viewer_status(&message.info.id),
-                    })
-                }
-                _ => None,
-            })
-            .collect::<Vec<_>>();
-        if !attachments
-            .iter()
-            .any(|item| item.message_id == selected.info.id)
-        {
-            attachments.push(crate::app::events::ViewerAttachment {
-                message_id: selected.info.id.clone(),
-                kind: file.kind.clone(),
-                path: file.path.clone(),
-                status: self.viewer_status(&selected.info.id),
-            });
-        }
-        let index = attachments
-            .iter()
-            .position(|item| item.message_id == selected.info.id)
-            .unwrap_or_default();
-        self.attachment_viewer = Some(crate::app::events::AttachmentViewerState::from_attachments(
-            attachments,
-            index,
-        ));
-        self.viewer_preview = None;
-    }
-
-    fn viewer_status(&self, message_id: &wr::MessageId) -> crate::app::events::ViewerStatus {
-        match self.metadata.get(message_id) {
-            Some(crate::app::Metadata::File(
-                crate::app::FileMeta::Downloaded | crate::app::FileMeta::Loaded,
-            )) => crate::app::events::ViewerStatus::Ready,
-            Some(crate::app::Metadata::File(
-                crate::app::FileMeta::Downloading | crate::app::FileMeta::Loading,
-            )) => crate::app::events::ViewerStatus::Downloading,
-            Some(crate::app::Metadata::File(
-                crate::app::FileMeta::DownloadFailed | crate::app::FileMeta::LoadFailed,
-            )) => crate::app::events::ViewerStatus::Failed,
-            _ => crate::app::events::ViewerStatus::Missing,
-        }
-    }
-
-    fn navigate_viewer(&mut self, delta: isize) {
-        if let Some(viewer) = &mut self.attachment_viewer {
-            viewer.navigate(delta);
-            self.viewer_preview = None;
-        }
-    }
-
-    fn plan_viewer_media_launch(&mut self) {
-        let Some(viewer) = self.attachment_viewer.as_ref().cloned() else {
-            return self.unavailable("Media is not downloaded");
-        };
-        if viewer.status != crate::app::events::ViewerStatus::Ready {
-            return self.unavailable("Media is not downloaded");
-        }
-        self.plan_media_launch(&viewer.kind, Path::new(viewer.path.as_ref()));
-    }
-
-    fn plan_selected_media_launch(&mut self) {
-        if self.selected_message_is_deleted() {
-            return self.unavailable("This message was deleted.");
-        }
-        let Some(message) = self.selected_message() else {
-            return self.unavailable("Media is not downloaded");
-        };
-        let wr::MessageContent::File(file) = &message.message else {
-            return self.unavailable("Media is not downloaded");
-        };
-        if !matches!(
-            self.metadata.get(&message.info.id),
-            Some(crate::app::Metadata::File(
-                crate::app::FileMeta::Downloaded | crate::app::FileMeta::Loaded
-            ))
-        ) {
-            return self.unavailable("Media is not downloaded");
-        }
-        let player = match crate::media::media_opener_from_environment(&file.kind) {
-            Ok(player) => player,
-            Err(_) => {
-                return self.unavailable(
-                    "Set WPTUI_IMAGE_VIEWER or WPTUI_MEDIA_PLAYER to one executable name",
-                );
-            }
-        };
-        match crate::media::plan_media_launch(
-            Some(&self.media_path),
-            Some(&player),
-            Some(Path::new(file.path.as_ref())),
-        ) {
-            Ok(Some(plan)) => {
-                match crate::media::execute_launch(&plan, &mut crate::media::CommandLaunchExecutor)
-                {
-                    Ok(()) => {
-                        self.action_notice = Some(crate::app::actions::ActionNotice::Unavailable(
-                            "Media player started".into(),
-                        ))
-                    }
-                    Err(error) => {
-                        self.unavailable(&format!("Could not start media player: {error:?}"))
-                    }
-                }
-            }
-            Ok(None) | Err(_) => self.unavailable("Media launch is unavailable"),
-        }
-    }
-
-    fn plan_media_launch(&mut self, kind: &wr::FileKind, path: &Path) {
-        let player = match crate::media::media_opener_from_environment(kind) {
-            Ok(player) => player,
-            Err(_) => {
-                return self.unavailable(
-                    "Set WPTUI_IMAGE_VIEWER or WPTUI_MEDIA_PLAYER to one executable name",
-                );
-            }
-        };
-        match crate::media::plan_media_launch(Some(&self.media_path), Some(&player), Some(path)) {
-            Ok(Some(plan)) => {
-                match crate::media::execute_launch(&plan, &mut crate::media::CommandLaunchExecutor)
-                {
-                    Ok(()) => {
-                        self.action_notice = Some(crate::app::actions::ActionNotice::Unavailable(
-                            "Media player started".into(),
-                        ))
-                    }
-                    Err(error) => {
-                        self.unavailable(&format!("Could not start media player: {error:?}"))
-                    }
-                }
-            }
-            Ok(None) | Err(_) => self.unavailable("Media launch is unavailable"),
-        }
     }
 
     fn copy_selected_text(&mut self) {
